Piece Class
Some musings on the piece superclass we can use, and an idea on having a .name instance variable for the board square objects. We can quickly determine where the piece is by the square it's on, and if the GUI coordinate system works well, the ord() stuff should be able to figure out what the Point of the piece is.